resume_processing_gui/requirements.txt
langchain-core
pydantic
langgraph
google-generativeai
python-docx
pypdf2
streamlit
python-dotenv